[CONFIG/#2] Skip 프로젝트 템플릿 정리
