+++ e-eum/Sources/E_Eum/Tmp/ItemView.swift
@@ -1,0 +1,34 @@
+import SwiftUI
+
+struct ItemView : View {
+    @State var item: Item
+    @Environment(ViewModel.self) var viewModel: ViewModel
+    @Environment(\.dismiss) var dismiss
+
+    var body: some View {
+        Form {
+            TextField("Title", text: $item.title)
+                .textFieldStyle(.roundedBorder)
+            Toggle("Favorite", isOn: $item.favorite)
+            DatePicker("Date", selection: $item.date)
+            Text("Notes").font(.title3)
+            TextEditor(text: $item.notes)
+                .border(Color.secondary, width: 1.0)
+        }
+        .navigationBarBackButtonHidden()
+        .toolbar {
+            ToolbarItem(placement: .cancellationAction) {
+                Button("Cancel") {
+                    dismiss()
+                }
+            }
+            ToolbarItem(placement: .confirmationAction) {
+                Button("Save") {
+                    viewModel.save(item: item)
+                    dismiss()
+                }
+                .disabled(!viewModel.isUpdated(item))
+            }
+        }
+    }
+}
