+++ e-eum/Sources/E_Eum/Tmp/PlatformHeartView.swift
@@ -1,0 +1,14 @@
+import SwiftUI
+
+/// A view that shows a blue heart on iOS and a green heart on Android.
+struct PlatformHeartView: View {
+    var body: some View {
+       #if SKIP
+       ComposeView { ctx in // Mix in Compose code!
+           androidx.compose.material3.Text("💚", modifier: ctx.modifier)
+       }
+       #else
+       Text(verbatim: "💙")
+       #endif
+    }
+}
